+++ front/app/modules/commercial/user_custom_fields/services/__mocks__/userCustomFieldsStats.ts
@@ -9,6 +9,11 @@ export const mockUsersByRegFields = {
       _blank: 11,
     },
     expected_users: {
+      active_politician: 0.9,
+      retired_politician: 0.4,
+      no: 8.7,
+    },
+    reference_population: {
       active_politician: 10,
       retired_politician: 5,
       no: 100,